+++ CompanyManagement/Logic/Classes/DepartmentLogic.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using CompanyManagement.Logic.Interfaces;
 using CompanyManagement.Models;
 using CompanyManagement.Repository;
+using Microsoft.EntityFrameworkCore.ChangeTracking.Internal;
 
 namespace CompanyManagement.Logic
 {
@@ -25,19 +26,34 @@ namespace CompanyManagement.Logic
             return await _departmentRepository.ReadAsync(id);
         }
 
-        public async Task UpdateAsync(Department department)
+        public async Task<bool> UpdateAsync(Department department)
         {
+            var entity =await _departmentRepository.ReadAsync(department.Id);
+            if(department==null)
+            {
+                return false;
+            }
             await _departmentRepository.UpdateAsync(department);
+            return true;
         }
 
-        public async Task DeleteAsync(int id)
+        public async Task<bool> DeleteAsync(int id)
         {
+            var entity = await _departmentRepository.ReadAsync(id);
+            if(entity==null)
+            {
+                return false;
+            }
             await _departmentRepository.DeleteAsync(id);
+            return true;
         }
 
         public async Task<IEnumerable<Department>> GetAllDepartmentsAsync()
         {
             return await _departmentRepository.ReadAllAsync();
         }
+
+
+    
     }
 }
